chore(ops): move mod error to its own file
